+++ wp/soulshape.earth/wp-content/themes/zerif-lite/js/zerif_customizer.js
@@ -1,21 +1,5 @@
 jQuery(document).ready(function() {
-
-	/* Upsells in customizer (Documentation link and Upgrade to PRO link */
-	if( !jQuery( ".zerif-upsells" ).length ) {
-		jQuery('#customize-theme-controls > ul').prepend('<li class="accordion-section zerif-upsells">');
-	}
-
-	if( jQuery( ".zerif-upsells" ).length ) {
-
-		jQuery('.zerif-upsells').append('<a style="width: 80%; margin: 5px auto 5px auto; display: block; text-align: center;" href="http://themeisle.com/documentation-zerif-lite" class="button" target="_blank">{documentation}</a>'.replace('{documentation}', zerifLiteCustomizerObject.documentation));
-
-	}
-	jQuery('.preview-notice').append('<a class="zerif-upgrade-to-pro-button" href="http://themeisle.com/themes/zerif-pro-one-page-wordpress-theme/" class="button" target="_blank">{pro}</a>'.replace('{pro}',zerifLiteCustomizerObject.pro));
-
-	if ( !jQuery( ".zerif-upsells" ).length ) {
-		jQuery('#customize-theme-controls > ul').prepend('</li>');
-	}
-
+	
 	jQuery( '.ui-state-default' ).on( 'mousedown', function() {
 		jQuery( '#customize-header-actions #save' ).trigger( 'click' );
 
@@ -50,5 +34,12 @@ jQuery(document).ready(function() {
 		jQuery(this).next('.zerif-moreinfo-content').show();
 	},function(){
 		jQuery(this).next('.zerif-moreinfo-content').hide();
+	});
+
+	//Locked sections
+	jQuery('#accordion-section-zerif_order_section, #accordion-section-zerif_videobackground_in_pro_section').click(function() {
+		jQuery('.wp-full-overlay').removeClass('section-open');
+		jQuery('#accordion-section-zerif_order_section, #accordion-section-zerif_videobackground_in_pro_section').removeClass('open');
+		window.location.href = "http://themeisle.com/themes/zerif-pro-one-page-wordpress-theme/";
 	});
 });