+++ wp/soulshape.earth/wp-content/themes/zerif-lite/CHANGELOG.md
@@ -1,4 +1,17 @@
 
+### 1.8.4.9 - 30/08/2016
+**Changes:** 
+- Use core function for archive title and description
+- Remove wp_title backward compatibility
+- Fixed issue with background image when using a child theme
+- Redo upsell section using Customizer API
+
+### 1.8.4.8 - 29/08/2016
+**Changes:** 
+- NEW option to change the Clients area title
+- Changed the way locked sections link to pro page
+- Add note in readme.txt regarding the background image
+
 ### 1.8.4.7 - 26/08/2016
 **Changes:** 
 - Replace back string for template name